Add a prop that changes the layout engine.
This is not working 100% for me, it only updates on resize but in general people will want to set it once so that is probably OK.

I also readded persitent props. This had been done earlier, but was lost in a delete when my environemnt broke.

I am gettign some dash errors that don;t seem to make sense, so  I am pushing this to a branch for people to look  into.